G2: Split TodayFeed JSON models and state classes into separate part file for component size governance
